Add simple control of RD unit.
SpinCtrlDoubles for set voltage and amperage, and a ToggleButton for
enabling the output.

Add a second field to the status bar, for the last read voltage and
amperage.

NOTE: there is unexpected behavior (bug?) with the SpinCtrlDouble, where
writing a number in the text box and hitting "enter" triggers the
default ("Update") button, but the edit in the text box is not applied
to the control's value.