Size tweaks.
The Frame size doesn't seem to have been used, so removed.

The FloatSpinCtrl's default width was too narrow on Windows, but on
Linux/GTK the SpinButton is HUGE, too big to fit in the space the whole
control takes on Windows.  Luckily, the default size on Linux/GTK was
big enough, so only set the size on platform="win".  TODO: check on OSX,
if anybody cares.