Turn on grid lines and minor ticks for y axes.
The default 'minor locator' would use either 4 or 5 minor ticks per
major tick, depending on the values of the major ticks.  It so happened
that the scales I was using for voltage and amperage worked out to
choose differently for the two Y axes, which was not nice, so override
to always use 4 minor ticks per major tick.